+++ src/features/favorite/index.ts
@@ -1,4 +1,3 @@
-import { editFavoritesMessage } from "../../eventListener";
 import { getCourseSiteID } from "../../utils";
 
 /**
@@ -86,3 +85,22 @@ export const addFavoritedCourseSites = (baseURL: string): Promise<void> => {
         request.send();
     });
 };
+
+async function editFavoritesMessage(): Promise<void> {
+    // Wait 200ms until jQuery finished generating message.
+    await new Promise((r) => setTimeout(r, 200));
+    try {
+        const message = document.getElementsByClassName("favorites-max-marker")[0];
+        // TODO: i18n
+        message.innerHTML = `<i class='fa fa-bell warning-icon'></i>${
+            chrome.runtime.getManifest().name
+        }によってお気に入り登録した<br>サイトが全てバーに追加されました。`;
+        const lectureTabs = document.getElementsByClassName("fav-sites-entry");
+        const lectureTabsCount = lectureTabs.length;
+        for (let i = 0; i < lectureTabsCount; i++) {
+            lectureTabs[i].classList.remove("site-favorite-is-past-max");
+        }
+    } catch (e) {
+        console.log("could not edit message");
+    }
+}
